refactor(resampling): make typing module internal
