+++ quantum_enablement/resampling/_typing.py
